Use sed instead of ed for libtool SONAME hacking
Since sed is already required by autoconf and libtool, it makes sense to
use it instead of pulling in ed, too. The editing script only has to be
changed slightly to accommodate this. The resulting libtool has been
checked to be the same on Linux (GNU sed) and HP-UX.

Building libXaw was the only reason I have ed on my system anymore.
Maybe someday this can be fixed properly.