Remove almost 10 year old notice about Xaw development from man page.
  Correct a problem when deleting text portions that may lead to a crash
due to accessing released memory; in most cases it would crash in
AsciiSink.c:AsciiPreparePaint() when calling XtMalloc() with a negative
value.
  The problem was due to  not properly removing/updating XawTextEntity
objects. One case that can cause the problem to happen is with patch files
with syntax highlight, and sequentially selecting a deleting them. At one
point attempt to remove the "first" XawTextAnchor, and also incorrectly
update it's offset value to a negative value. The problem is not easy
to reproduce as it requires a file with a special pattern (in this case
the syntax highlight used for patch) and it would fail after removing
all "entities" of the second "anchor".