Avoid warning about missing getpagesize() prototype on Windows
Work-around AC_CHECK_FUNCS([getpagesize]) reporting a false positive
when using MinGW gcc, due to it being present in libgcc.a, which results
in trying to use it without a prototype.

Future work: This value control the default memory allocation size for
text widgets.  If there's some reason why a fallback to using BUFSIZ is
bad, we could use the actual pagesize instead.